+++ src/main/java/br/com/blog/modules/comment/services/CommentService.java
@@ -12,6 +12,8 @@ import br.com.blog.modules.comment.repositories.CommentRepository;
 @Service
 public class CommentService {
     
+    private final String MESSAGE_DEFAULT_COMMENT = "This user does not have permission to delete this comment";
+    
     @Autowired
     private CommentRepository commentRepository;
     
@@ -24,8 +26,16 @@ public class CommentService {
         return comment;
     }
     
-    public void deleteCommentById(Long id) {
+    public void deleteCommentById(Long idUser, Long id) {
+
+        Optional<Comment> comment = commentRepository.findById(id);
+
+        if (!comment.get().getIdUser().equals(idUser)) {
+            throw new IllegalArgumentException(MESSAGE_DEFAULT_COMMENT);
+        }
+        
         commentRepository.deleteById(id);
+
     }
     
     public Optional<Comment> findByIdComment(Long id) {
